+++ src/main/java/insane96mcp/iguanatweaksreborn/mixin/plugin/ModLoadedPlugin.java
@@ -27,6 +27,7 @@ public class ModLoadedPlugin implements IMixinConfigPlugin {
             //.put("insane96mcp.iguanatweaksreborn.mixin.integration.farmersdelight.PieBlockMixin", () -> LoadingModList.get().getModFileById("farmersdelight") != null)
             .put("insane96mcp.iguanatweaksreborn.mixin.integration.cavernsandchasms.CCEntityTypesMixin", () -> LoadingModList.get().getModFileById("caverns_and_chasms") != null)
             .put("insane96mcp.iguanatweaksreborn.mixin.integration.appleskin.HUDOverlayHandlerMixin", () -> LoadingModList.get().getModFileById("appleskin") != null)
+            .put("insane96mcp.iguanatweaksreborn.mixin.integration.villagercomfort.ComfortHelperMixin", () -> LoadingModList.get().getModFileById("villagercomfort") != null)
             .build();
     @Override
     public void onLoad(String mixinPackage) {